Module scaling_up_staying_secure::agent

Defines the agents that are part of the model and the trait that is used to create new ones.

To take part in the simulation agents need to implement the AgentTrait trait. This allows them to be a part of the underlying graph and take part in the grid modelling.

Furthermore, some types are defined here that make it easier to read the code. These are nothing else than other names for data structures.

New Agents

Implementing new agents can be done by implementing AgentTrait and adding the discriminant to the AgentKind enum. The behaviour during each step of the model needs to be defined in the model part.

Structs

Area

The Area agent. A level below Root and above Netstation.

Household

The Household agent. A level below Netstation and lowest on the grid.

Netstation

The Netstation agent. A level below Area and one above Household.

Root

The Root agent. Top-level of the grid and sits above Area.

Enums

AgentKind

Agents and results in better readability of the code at the small expense of enum

Traits

AgentTrait

Implementing this trait lets structs participate in the model. New agents need to implement this trait.

Type Definitions

AgentList

A vector of the AgentRef type.

AgentRef

This type can be used to indicate a struct that implements the AgentTrait within a RwLock and Arc.

AreaList

A vector of the AreaRef type.

AreaRef

The Area struct within an Arc and RwLock.

HouseholdList

A vector of the HouseholdRef type.

HouseholdRef

The Household struct within an Arc and RwLock.

NetstationList

A vector of the NetstationRef type.

NetstationRef

The Netstation struct within an Arc and RwLock.