Struct scaling_up_staying_secure::agent::Netstation

pub struct Netstation {
    pub kind: AgentKind,
    pub index: BaseUint,
    pub step: Steps,
    pub children: AgentList,
    pub powerstate: PowerState,
    pub volt_boundary: Boundaries<mVolt>,
    pub volt_state: VoltState,
}

The Netstation agent. A level below Area and one above Household.

A Netstation is a power distribution house that is situated in a neighbourhood. They take the High Voltage of the grid and transform this to the known 230V. As this is the point where the voltage of the grid is determined, this agent is responsible for the Voltage stability monitoring of the grid.

Fields
kind: AgentKind

Kind of agent.

index: BaseUint

Index used in the graph.

step: Steps

Current step of the model.

children: AgentList

List of children this Agent has according to the graph.

powerstate: PowerState

The current power state of the agent. Changes each step.

volt_boundary: Boundaries<mVolt>

Voltage boundary of the grid, that needs to be monitored at this agent.

volt_state: VoltState

The current voltage and history of voltage within the agent.

Implementations

impl Netstation

pub fn volt_state(&mut self) -> &mut VoltState

Gives mutable access to the volt_state field.

pub fn volt_boundary(&mut self) -> &mut Boundaries<mVolt>

Gives mutable access to the volt_boundary field.

pub fn new(index: BaseUint, grid_param: &Grid) -> Self

Creates a new netstation agent, given the grid parameters and index of the graph.

pub fn get_infection_state_children(&self) -> InfectionStatistics

Get the infection statistics for the children of the netstation