Trait scaling_up_staying_secure::agent::AgentTrait

pub trait AgentTrait: Debug {

    fn kind(&self) -> &AgentKind;
    fn index(&self) -> &BaseUint;
    fn step(&self) -> &Steps;
    fn powerstate(&self) -> &PowerState;
    fn children(&self) -> &AgentList;
    fn kind_mut(&mut self) -> &mut AgentKind;
    fn index_mut(&mut self) -> &mut BaseUint;
    fn step_mut(&mut self) -> &mut Steps;
    fn children_mut(&mut self) -> &mut AgentList;
    fn powerstate_mut(&mut self) -> &mut PowerState;
    fn update_step(&mut self, step: Steps);
    fn get_json(&self) -> Result<Vec<u8>, ModelError>;
    fn get_avro(&self) -> Result<(Schema, Value), ModelError>;

    fn power_gen(&self) -> Option<&PowerGeneration> { ... }
    fn power_gen_mut(&mut self) -> Option<&mut PowerGeneration> { ... }
    fn calc_power_from_child(&mut self) { ... }

}

Implementing this trait lets structs participate in the model. New agents need to implement this trait.

The trait gives access to the inner data of the Agent, therefore this trait forces also that the Agent structs have the correct inner fields. Other methods that you want to use on the agents can be implemented on the struct themselves.

Required Methods

fn kind(&self) -> &AgentKind

Read access to kind field.

fn index(&self) -> &BaseUint

Read access to index field.

fn step(&self) -> &Steps

Read access to step field.

fn powerstate(&self) -> &PowerState

Read access to powerstate field.

fn children(&self) -> &AgentList

Read access to children field.

fn kind_mut(&mut self) -> &mut AgentKind

Mutuable access to the kind field.

fn index_mut(&mut self) -> &mut BaseUint

Mutuable access to the index field.

fn step_mut(&mut self) -> &mut Steps

Mutable access to the step field.

fn children_mut(&mut self) -> &mut AgentList

Mutable access to the chlidren field.

fn powerstate_mut(&mut self) -> &mut PowerState

Mutable access to the powerstate field.

fn update_step(&mut self, step: Steps)

Set the step field of the agent.

fn get_json(&self) -> Result<Vec<u8>, ModelError>

Transform the agent to a JSON representation of itself.

fn get_avro(&self) -> Result<(Schema, Value), ModelError>

Transform the agent to an AVRO representation of itself.

Provided Methods

fn power_gen(&self) -> Option<&PowerGeneration>

Read access to power_generation field.

fn power_gen_mut(&mut self) -> Option<&mut PowerGeneration>

Mutable access to power_generation field.

fn calc_power_from_child(&mut self)

This method is called during a step of the model to determine the powerstate of the Agent if it has children.