Module scaling_up_staying_secure::grid

All things related to the grid and power can be found here.

A few different topics are part of this module.

Construction of the Grid

The construction of the gird is done by the parameters given in the GridParameters file that was given to the executable by the ModelParameters file.

Parameters

Bulk consumption.
The amount of Watt that is consumed by big energy consumers. The bigger this number the lower the instability of the grid.

Boundary values.
For both frequency and voltage, the allowed limits are given.
These are extracted from EU regulations.

Number of agents per layer.
How many agents are needed per layer? Currently, no connections are made on a horizontal level.
So this is the number of children per node.

PV adoption.
The number of households that have a PV installation at the home.
This may or may not be infectable.

Power generation

Generation of power is done in this model by PV installations. However, this is a subset of Distributed Energy Resources (DER) that can be used in its place instead.

Generation characteristics are based on actual household data given by the MFFBAS document. PV households are assumed to be net zero throughout the year. But do not have energy storage in the house. To let each house be unique the standardized profile is adjusted with noise functions and a linear modifier. This gives dynamic in the grid.

Power mismatch

When there is too much or too little power being generated with regard to the power being consumed this has an impact on the voltage and frequency of the grid. This is monitored by the Root and Netstation agents.

Structs

Boundaries

A struct describing the boundaries that are set for type T. If these are exceeded they will be logged. Currently intended for Voltage and Frequency on the grid.

BoundaryBand

Restriction on the grid. How long may the grid pass the border in the number of steps? Generated from the GridParameters file.

FreqState

Frequency state of an agent. With frequency, this is only the root node.

Grid

Struct that defines the grid parameters

GridWarning

The warning returned if the grid is outside normal parameters.

InfectionCount

A helper struct to determine the state of the grid.

InfectionStatistics

A helper struct to determine the statistics on the infection in the grid.

NormalBand

What is considered normal behaviour of this parameter? Given by the GridParameters file.

PowerGeneration

Contains the power generation information for a household.

PowerState

A struct containing the power state of an agent.

ReservePower

Grid energy storage, that can be used to regulate.

SineParam

Contains the parameters of the sine function that create the behaviour. Note, the b attribute goes from 0 to 1 for a whole period. y = a * sin(bx + c) + d

VoltState

Voltage state of an agent. With voltage, these are the Netstation agents as they transform the voltage to the familiar 230V.

Enums

GridBoundaryState

The state the grid is in.

InfectionState

All possible states that a PV system can have regarding being infected.

Traits

BoundaryAgentTrait

A helper trait to enable checking the boundaries

BoundaryUnitTrait

Helper trait, to implement on the units used in the boundary checking.

GridState

A trait to be implemented for the grid states. Gives an interface to work with.