Struct scaling_up_staying_secure::model::Model

pub struct Model {
    pub param: ModelParameters,
    pub graph: MyGraph,
    pub agents: AgentList,
    pub reserve_power: ReservePower,
    pub root: Arc<RwLock<Root>>,
    pub areas: Vec<Arc<RwLock<Area>>>,
    pub netstations: Vec<Arc<RwLock<Netstation>>>,
    pub households: Vec<Arc<RwLock<Household>>>,
}

Struct for creating and running a model based on the respective ModelParameters.

Fields
param: ModelParameters

The parameters of the model. See ModelParameters

graph: MyGraph

The graph of the model

agents: AgentList

The list of agents in the model.

reserve_power: ReservePower

ReservePower

root: Arc<RwLock<Root>>

The Root agent of the model.

areas: Vec<Arc<RwLock<Area>>>

A list of Area agents that are part of the model

netstations: Vec<Arc<RwLock<Netstation>>>

A list of Netstation agents that are part of the model.

households: Vec<Arc<RwLock<Household>>>

A list of Household agents that are part of the model.

Implementations

impl<'a> Model

pub fn step(&'a mut self, num_steps: Steps) -> Result<(), ModelError>

Take the next step in the model. A step is also called a tick in some literature.

pub fn update_step(agents: &AgentList, step: Steps)

Update step

pub fn subagent(agentkind: AgentKind, agents: &AgentList) -> AgentList

Create a vector that copies the Arc of the agents with the desired AgentKind.

pub fn grid_frequency_compensation(&mut self)

Try to compensate for the power mismatch on the Root node.

pub fn update_history(agents: &AgentList)

Update the history of the agents given.

impl Model

pub fn new(model_param: ModelParameters) -> Result<Self, ModelError>

Creating the graph and agents and adding agents to the graph. Graph is tree. Ordinality: Area -> Netstation -> Households -> PV.

pub fn run(&mut self) -> Result<(), ModelError>

Run the model. Generates the outcome by calling the step function.

impl Model

pub fn output_step_agents_avro(agents: &AgentList) -> Vec<(Schema, AvroValue)>

Output the state of the model in avro format by transforming each agent into an AvroValue.

pub fn output_step_model_avro(
param: &ModelParameters,
step: Steps,
agent_values: Vec<(Schema, AvroValue)>,
handles: &mut Vec<JoinHandle<()>>
)

Log the state of the model in this step.

impl Model

pub fn output_step_agents_json(agents: &AgentList) -> Vec<Vec<u8>>

Transforms the vector of agents to a vector of JSON values in bytes.

pub fn output_step_model_json(
param: &ModelParameters,
step: Steps,
agent_values: Vec<Vec<u8>>,
handles: &mut Vec<JoinHandle<()>>
)

Log the state of the model in this step.

impl Model

pub fn output(
&self,
step: Steps,
handles: &mut Vec<JoinHandle<()>>
) -> Result<(), ModelError>

Output the model in its current state.