Struct scaling_up_staying_secure::model::ModelParameters

pub struct ModelParameters {

    pub name: String,
    pub steps: Steps,
    pub seed: SmallRng,
    pub logfolder: String,
    pub logfile: String,
    pub enable_output: bool,
    pub type_output: String,
    pub outputdatafolder: String,
    pub outputdatafile: String,
    pub outputcodec: Codec,
    pub attack: Attack,
    pub grid: Grid,
    pub loglevel: LevelFilter,
    pub stop_on_freq_error: bool,

}

Struct that defines the model parameters

Fields
name: String

Naming for the model. Used for logs etc.

steps: Steps

Number of steps in the Model

seed: SmallRng

Seed for the random generators, same behaviour between runs

logfolder: String

Folder to store logs

logfile: String

Chrono - format string to determine the filename. link

enable_output: bool

Enable the output of the node states per step.

type_output: String

Type of output, avro or json.

outputdatafolder: String

Folder to store the output

outputdatafile: String

Chrono - format string to determine the filename. link

outputcodec: Codec

The codec that is used. Currently supported: “snappy”, “deflate”, ’null“, See also [Codec].

attack: Attack

The attack struct of this model. See Attack.

grid: Grid

The Grid struct of this model. See Grid.

loglevel: LevelFilter

The log level of the executable.

stop_on_freq_error: bool

Stop running the model when there is a grid frequency error.

Implementations

impl ModelParameters

pub fn test() -> Self

Creates a test version to be used for testing within the crate.

impl ModelParameters

pub fn from_config(filepath: impl AsRef<Path>) -> Result<Self, ConfigError>

Create the ModelParameters from a “ModelParameters.ini” file in the directory.